fix: Change int to BigDecimal

+++ src/main/java/mg/sandratra/bakery/dto/recipe/RecipeIngredientDto.java
@@ -1,5 +1,7 @@
 package mg.sandratra.bakery.dto.recipe;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,6 +11,6 @@ import mg.sandratra.bakery.models.ingredient.Ingredient;
 @NoArgsConstructor
 @AllArgsConstructor
 public class RecipeIngredientDto {
-    private int quantity;
+    private BigDecimal quantity;
     private Ingredient ingredient;
 }
